% Auto-generated: do not edit by hand
\name{muiButton}

\alias{muiButton}

\title{Button component}

\description{
Button wraps Material-UI Button.  Use the `n_clicks` prop to trigger callbacks when the button has been clicked.
}

\usage{
muiButton(id=NULL, text=NULL, variant=NULL, color=NULL, href=NULL,
disableElevation=NULL, disabled=NULL, n_clicks=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{text}{Character. Button text.}

\item{variant}{Character. Material-UI Button variant prop.}

\item{color}{Character. Material-UI Button color prop.}

\item{href}{Character. Material-UI Button href prop.}

\item{disableElevation}{Logical. Material-UI Button disableElevation prop.}

\item{disabled}{Logical. Material-UI Button disabled prop.}

\item{n_clicks}{Numeric. An integer that represents the number of times that this element has been clicked on.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

