% Auto-generated: do not edit by hand
\name{muiTable}

\alias{muiTable}

\title{Table component}

\description{
Table wraps Material-UI Table.
}

\usage{
muiTable(id=NULL, title=NULL, rowColor=NULL, headerColor=NULL,
data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{title}{Character. Material Table title}

\item{rowColor}{Character. Material Table Row Color}

\item{headerColor}{Character. Material Table Header Color}

\item{data}{Unnamed list. Material Table Data}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

